Refactor: patient and chat types to shared types directory
Moved Patient and ChatMessage type definitions to a new 'types' directory for better code organization and reusability. Updated all relevant imports in doctor components to reference the new type locations.
