+++ frontend/src/component/doctor/Calender.tsx
@@ -2,7 +2,7 @@ import { useState } from "react"
 import { Calendar, Star, User } from "lucide-react"
 import { useNavigate } from "react-router-dom"
 import { Calendar as CalendarIcon } from "../../components/ui/calendar"
-import { dummyPatients } from "../Patient"
+import { dummyPatients } from "../../types/Patient"
 import { Card, CardContent } from "../../components/ui/card"
 import { format } from "date-fns"
 import { ko } from "date-fns/locale"
