+++ frontend/src/types/ChatMessage.tsx
@@ -1,0 +1,4 @@
+export interface ChatMessage {
+  question: string;
+  answer: string;
+}
